pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llms_orchestration"
version = "0.0.1"
description = "A multi-agent system for LLM orchestration course project"
authors = [
  { name="Itamar V", email="itamarv@example.com" },
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pypdf",
    "pytest",
    "matplotlib",
    "seaborn",
    "jupyter",
    "ruff",
    "bandit",
    "python-dotenv",
    "google-generativeai",
    "pillow"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "mypy"
]

[tool.setuptools.packages.find]
where = ["src"]
